+++ src/base/b-dummy/interface.ts
@@ -8,10 +8,15 @@
 
 import { ImageLoader, imageLoaderFactory } from 'core/dom/image';
 import { InViewAdapter } from 'core/dom/in-view';
+import { ResizeWatcher } from 'core/dom/resize-observer';
 
 export interface Directives {
 	imageFactory: typeof imageLoaderFactory;
 	image: typeof ImageLoader;
 	inViewMutation: InViewAdapter;
 	inViewObserver: InViewAdapter;
+}
+
+export interface Modules {
+	resizeWatcher: typeof ResizeWatcher;
 }