+++ src/base/b-virtual-scroll/modules/chunk-request.ts
@@ -350,6 +350,7 @@ export default class ChunkRequest extends Friend {
 
 		return ctx.async.request(ctx.getData(this.component, params), {label: $$.request})
 			.then((data) => {
+				this.ctx.localState = 'ready';
 				void ctx.removeMod('progress', true);
 				this.lastLoadedChunk.raw = data;
 
@@ -365,11 +366,11 @@ export default class ChunkRequest extends Friend {
 			})
 
 			.catch((err) => {
+				void ctx.removeMod('progress', true);
+
 				if (isAsyncClearError(err)) {
 					return Promise.reject(err);
 				}
-
-				void ctx.removeMod('progress', true);
 
 				chunkRender.setRefVisibility('retry', true);
 				chunkRender.setRefVisibility('renderNext', false);