+++ src/base/b-virtual-scroll/test/runners/render/render.js
@@ -68,6 +68,7 @@ module.exports = (page) => {
 			];
 
 			globalThis.renderComponents('b-virtual-scroll', scheme);
+			globalThis.componentNode = document.querySelector('.b-virtual-scroll');
 		});
 
 		await h.bom.waitForIdleCallback(page);
@@ -87,7 +88,7 @@ module.exports = (page) => {
 					await component.evaluate((ctx) => ctx.request = {get: {chunkSize: 10, total: 0}});
 					await h.dom.waitForEl(container, 'section', {to: 'unmount'});
 
-					expect(await getContainerChildCount()).toBe(0);
+					expect(await component.evaluate((ctx) => ctx.$refs.container.childElementCount === 0)).toBeTrue();
 				});
 
 				it('renders new', async () => {
@@ -96,6 +97,7 @@ module.exports = (page) => {
 					const
 						chunkSize = await component.evaluate((ctx) => ctx.requestParams.get.chunkSize);
 
+					await h.dom.waitForEl(container, `section:nth-child(${chunkSize - 1})`);
 					expect(await getContainerChildCount()).toBe(chunkSize);
 
 					await component.evaluate((ctx) => ctx.request = {get: {chunkSize: 4, total: 4, id: 'uniq-options'}});
@@ -106,6 +108,7 @@ module.exports = (page) => {
 					const
 						newChunkSize = await component.evaluate((ctx) => ctx.requestParams.get.chunkSize);
 
+					await h.dom.waitForEl(container, `section:nth-child(${newChunkSize - 1})`);
 					expect(await getContainerChildCount()).toBe(newChunkSize);
 				});
 			});
@@ -138,12 +141,12 @@ module.exports = (page) => {
 					await component.evaluate((ctx) => ctx.dataProvider = undefined);
 					await h.dom.waitForEl(container, 'section', {to: 'unmount'});
 
-					expect(await getContainerChildCount()).toBe(0);
+					expect(await component.evaluate((ctx) => ctx.$refs.container.childElementCount === 0)).toBeTrue();
 				});
 
 				it('renders new', async () => {
 					await h.dom.waitForEl(container, 'section', {to: 'unmount'});
-					expect(await getContainerChildCount()).toBe(0);
+					expect(await component.evaluate((ctx) => ctx.$refs.container.childElementCount === 0)).toBeTrue();
 
 					await component.evaluate((ctx) => ctx.dataProvider = 'demo.Pagination');
 					await h.dom.waitForEl(container, 'section');
@@ -275,9 +278,9 @@ module.exports = (page) => {
 
 		describe('without `options` and` dataProvider` specified', () => {
 			it('does not render anything', async () => {
-				expect(await component.evaluate((ctx) => ctx.options.length)).toBe(0);
-				expect(await component.evaluate((ctx) => ctx.dataProvider)).toBeUndefined();
-				expect(await getContainerChildCount()).toBe(0);
+				expect(await component.evaluate((ctx) => ctx.options.length === 0)).toBeTrue();
+				expect(await component.evaluate((ctx) => ctx.dataProvider === undefined)).toBeTrue();
+				expect(await component.evaluate((ctx) => ctx.$refs.container.childElementCount === 0)).toBeTrue();
 			});
 		});
 	});