+++ src/core/dom/image/test/index.js
@@ -35,6 +35,11 @@ module.exports = async (page, params) => {
 		imgNode,
 		divNode;
 
+	let
+		isClosed = false;
+
+	page.on('close', () => isClosed = true);
+
 	const handleImageRequest = (url, sleep = 0, base64Img = images.pngImage) => page.route(url, async (route) => {
 		await delay(sleep);
 
@@ -43,11 +48,15 @@ module.exports = async (page, params) => {
 			return;
 		}
 
-		const
+	const
 			res = base64Img.split(',')[1],
 			headers = route.request().headers();
 
 		headers['Content-Length'] = String(res?.length ?? 0);
+
+		if (isClosed) {
+			return;
+		}
 
 		route.fulfill({
 			status: 200,
@@ -474,6 +483,7 @@ module.exports = async (page, params) => {
 						});
 					}, [tag, images.pngImage, images.webp]);
 
+					// @ts-expect-error
 					await expectAsync(waitFor(getNode(tag), (ctx) => globalThis.getSrc(ctx) === document.getElementById('expected-img').currentSrc)).toBeResolved();
 				});
 
@@ -922,7 +932,7 @@ module.exports = async (page, params) => {
 
 				imageLoaderCtx.init(target, {
 					src: mainSrc,
-					bgOptions: {size: 'contain', ratio: 100 / 50, beforeImg, afterImg, position: '47% 47%'},
+					bgOptions: {size: 'contain', ratio: 100 / 50, beforeImg, afterImg, position: '47% 47%', repeat: 'no-repeat'},
 					ctx: globalThis.dummy
 				});
 			}, [tag, images.pngImage, beforeImg, afterImg]);
@@ -941,6 +951,7 @@ module.exports = async (page, params) => {
 			expect(await getNode(tag).evaluate((ctx) => ctx.style.paddingBottom)).toBe(expected);
 			expect(await getNode(tag).evaluate((ctx) => ctx.style.backgroundPosition)).toBe('47% 47%');
 			expect(await getNode(tag).evaluate((ctx) => ctx.style.backgroundSize)).toBe('contain');
+			expect(await getNode(tag).evaluate((ctx) => ctx.style.backgroundRepeat)).toBe('no-repeat');
 		});
 
 		it('div main with `src`, default ratio', async () => {
@@ -958,6 +969,7 @@ module.exports = async (page, params) => {
 				const testImg = document.createElement('img');
 				testImg.src = mainSrc;
 
+				// @ts-expect-error
 				testImg.onInit(() => {
 					if (testImg.naturalHeight > 0 || testImg.naturalWidth > 0) {
 						const ratio = testImg.naturalHeight === 0 ? 1 : testImg.naturalWidth / testImg.naturalHeight;
@@ -998,6 +1010,29 @@ module.exports = async (page, params) => {
 			expect(await divNode.evaluate((ctx) => globalThis.getSrc(ctx))).toBe(images.pngImage);
 			expect(await divNode.getAttribute('aria-label')).toBe('alt-text');
 			expect(await divNode.getAttribute('role')).toBe('img');
+		});
+
+		it('div tag initial padding bottom', async () => {
+			const
+				mainSrcUrl = getRandomImgUrl();
+
+			handleImageRequest(mainSrcUrl, 2000);
+
+			await imageLoader.evaluate((imageLoaderCtx, mainSrcUrl) => {
+				const div = document.getElementById('div-target');
+
+				imageLoaderCtx.init(div, {
+					src: mainSrcUrl,
+					bgOptions: {
+						ratio: 328 / 172
+					},
+					ctx: globalThis.dummy
+				});
+
+			}, mainSrcUrl);
+
+			await h.bom.waitForIdleCallback(page);
+			expect(await divNode.evaluate((ctx) => parseInt(ctx.style.paddingBottom, 10))).toBe(52);
 		});
 	});
 };