+++ src/form/b-button/b-button.ts
@@ -61,7 +61,7 @@ export default class bButton extends iData implements iAccess, iOpenToggle, iVis
 	readonly dataProvider: string = 'Provider';
 
 	/** @override */
-	readonly requestFilter: RequestFilter = false;
+	readonly defaultRequestFilter: RequestFilter = true;
 
 	/**
 	 * Link href
@@ -118,10 +118,11 @@ export default class bButton extends iData implements iAccess, iOpenToggle, iVis
 	readonly iconComponent?: string;
 
 	/**
-	 * Component for a progress icon
+	 * Component for a progress icon or
+	 * Boolean, if need to show progress by slot or `b-progress-icon`
 	 */
-	@prop({type: String, required: false})
-	readonly progressIcon?: string;
+	@prop({type: [String, Boolean], required: false})
+	readonly progressIcon?: string | boolean;
 
 	/**
 	 * Tooltip text