+++ src/base/b-virtual-scroll/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.81 (2020-10-08)
+
+#### :bug: Bug Fix
+
+* Fixed an issue with `renderNext`: hasn't been data rendering after a loading error
+
+## v3.0.0-rc.74 (2020-10-06)
+
+#### :bug: Bug Fix
+
+* Fixed an issue with removing the progress modifier
+
 ## v3.0.0-rc.68 (2020-09-23)
 
 #### :bug: Bug Fix