+++ src/core/component/construct/CHANGELOG.md
@@ -9,6 +9,22 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.84 (2020-10-09)
+
+#### :bug: Bug Fix
+
+* Fixed a bug when using a complex path as a dependency
+
+#### :house: Internal
+
+* Optimized creation of components
+
+## v3.0.0-rc.80 (2020-10-08)
+
+#### :bug: Bug Fix
+
+* Fixed a bug when using an array of dependencies to watch an accessor
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :house: Internal