+++ src/core/prelude/test-env/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.85 (2020-10-09)
+
+#### :house: Internal
+
+* Extracted interfaces to `index.d.ts`
+
 ## v3.0.0-rc.60 (2020-09-01)
 
 #### :rocket: New Feature